requirements.txt
opencv-python-headless
opencv-contrib-python-headless
Flask==1.1.1
gunicorn==20.0.4
Pillow
numpy

